basic metronome and foreground service

+++ cookietechmetronomelibrary/src/main/java/com/blz/cookietech/Listener/BPMListener.java
@@ -1,5 +1,5 @@
 package com.blz.cookietech.Listener;
 
 public interface BPMListener {
-    void onBPMChange(float bpm);
+    void onBPMChange(int bpm);
 }
